+++ src/CMakeLists.txt
@@ -1,8 +1,6 @@
 # Copyright (c) The ThingSet Project Contributors
 # SPDX-License-Identifier: Apache-2.0
 
-target_sources_ifdef(CONFIG_ISOTP_FAST app PRIVATE isotp_fast.c)
-
 target_sources_ifdef(CONFIG_THINGSET_SDK app PRIVATE sdk.c)
 target_sources_ifdef(CONFIG_THINGSET_SDK app PRIVATE packetizer.c)
