+++ subsys/canbus/isotp_fast/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (c) The ThingSet Project Contributors
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_sources(isotp_fast.c)
